+++ src/packages/piral-core/src/components/PiralRoutes.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import * as hooks from '../hooks';
 import { MemoryRouter } from 'react-router-dom';
 import { mount } from 'enzyme';
-import { Routes } from './Routes';
+import { PiralRoutes } from './PiralRoutes';
 
 const mountWithRouter = (node, url = '/') =>
   mount(
@@ -52,7 +52,7 @@ StubBarPage.displayName = 'StubBarPage';
 
 describe('Routes Module', () => {
   it('always goes to the given home on "/"', () => {
-    const node = mountWithRouter(<Routes NotFound={StubNotFound} />, '/');
+    const node = mountWithRouter(<PiralRoutes NotFound={StubNotFound} />, '/');
     expect(node.find(StubHome).length).toBe(1);
     expect(node.find(StubNotFound).length).toBe(0);
     expect(node.find(StubCustomPage).length).toBe(0);
@@ -60,7 +60,7 @@ describe('Routes Module', () => {
   });
 
   it('goes to the not found on an invalid path', () => {
-    const node = mountWithRouter(<Routes NotFound={StubNotFound} />, '/qxz');
+    const node = mountWithRouter(<PiralRoutes NotFound={StubNotFound} />, '/qxz');
     expect(node.find(StubHome).length).toBe(0);
     expect(node.find(StubNotFound).length).toBe(1);
     expect(node.find(StubCustomPage).length).toBe(0);
@@ -68,7 +68,7 @@ describe('Routes Module', () => {
   });
 
   it('goes to the custom page on "/custom"', () => {
-    const node = mountWithRouter(<Routes NotFound={StubNotFound} />, '/custom');
+    const node = mountWithRouter(<PiralRoutes NotFound={StubNotFound} />, '/custom');
     expect(node.find(StubHome).length).toBe(0);
     expect(node.find(StubNotFound).length).toBe(0);
     expect(node.find(StubCustomPage).length).toBe(1);
@@ -76,7 +76,7 @@ describe('Routes Module', () => {
   });
 
   it('goes exactly to the page on "/foo/bar"', () => {
-    const node = mountWithRouter(<Routes NotFound={StubNotFound} />, '/foo/bar');
+    const node = mountWithRouter(<PiralRoutes NotFound={StubNotFound} />, '/foo/bar');
     expect(node.find(StubHome).length).toBe(0);
     expect(node.find(StubNotFound).length).toBe(0);
     expect(node.find(StubCustomPage).length).toBe(0);
@@ -84,7 +84,7 @@ describe('Routes Module', () => {
   });
 
   it('goes exactly to the page on "/foo"', () => {
-    const node = mountWithRouter(<Routes NotFound={StubNotFound} />, '/foo');
+    const node = mountWithRouter(<PiralRoutes NotFound={StubNotFound} />, '/foo');
     expect(node.find(StubHome).length).toBe(0);
     expect(node.find(StubNotFound).length).toBe(0);
     expect(node.find(StubCustomPage).length).toBe(0);
@@ -93,7 +93,7 @@ describe('Routes Module', () => {
   });
 
   it('goes exactly to the page on "/bar"', () => {
-    const node = mountWithRouter(<Routes NotFound={StubNotFound} />, '/bar');
+    const node = mountWithRouter(<PiralRoutes NotFound={StubNotFound} />, '/bar');
     expect(node.find(StubHome).length).toBe(0);
     expect(node.find(StubNotFound).length).toBe(0);
     expect(node.find(StubCustomPage).length).toBe(0);
