+++ src/packages/piral-core/src/components/ResponsiveLayout.tsx
@@ -3,11 +3,11 @@ import { useMedia, useGlobalState, useAction } from '../hooks';
 import { defaultLayouts, defaultRender, defaultBreakpoints } from '../utils';
 import { LayoutBreakpoints } from '../types';
 
-export interface ResponsiveProps {
+export interface ResponsiveLayoutProps {
   breakpoints?: LayoutBreakpoints;
 }
 
-export const Responsive: React.FC<ResponsiveProps> = ({ breakpoints = defaultBreakpoints, children }) => {
+export const ResponsiveLayout: React.FC<ResponsiveLayoutProps> = ({ breakpoints = defaultBreakpoints, children }) => {
   const current = useGlobalState(m => m.app.layout) || 'desktop';
   const changeTo = useAction('changeLayout');
   const selected = useMedia(breakpoints, defaultLayouts, current);
@@ -18,4 +18,4 @@ export const Responsive: React.FC<ResponsiveProps> = ({ breakpoints = defaultBre
 
   return defaultRender(children);
 };
-Responsive.displayName = 'Responsive';
+ResponsiveLayout.displayName = 'ResponsiveLayout';
