+++ src/packages/piral-core/src/components/index.ts
@@ -5,9 +5,9 @@ export * from './DefaultRouter';
 export * from './ExtensionSlot';
 export * from './helpers';
 export * from './Mediator';
+export * from './PiralRoutes';
 export * from './PiralView';
-export * from './Responsive';
-export * from './Routes';
+export * from './ResponsiveLayout';
 export * from './SetComponent';
 export * from './SetRoute';
 export * from './withApi';
